Context API를 이용하여 여러차례 일어나는 전달 방지
