+++ Front/react_begin/src/UserList.js
@@ -1,6 +1,9 @@
-import React, {useEffect} from 'react';
+import React, {useEffect, useContext} from 'react';
+import { UserDispatch } from './App';
 
-const User = React.memo(function User( {user, onRemove, onToggle} ) {
+const User = React.memo(function User( {user} ) {
+  const dispatch = useContext(UserDispatch);
+
   useEffect(() => {
     console.log('user 값이 설정됨');
     console.log(user);
@@ -17,23 +20,27 @@ const User = React.memo(function User( {user, onRemove, onToggle} ) {
           cursor: 'pointer',
           color: user.active ? 'green' : 'black'
         }}
-        onClick={() => onToggle(user.id)}
+        onClick={() => {
+          dispatch({ type: 'TOGGLE_USER', id: user.id });
+        }}
       >
         {user.username}
       </b>
       &nbsp;
       {/* 위의 코드는 현 위치에서 자동 줄 바꿈(워드랩)을 막는 데 쓰인다. */}
       <span>({user.email})</span>
-      <button onClick={() => onRemove(user.id)}>삭제</button>
+      <button onClick={() => {
+        dispatch({ type: 'REMOVE_USER', id: user.id });
+      }}>삭제</button>
     </div>
   );
 });
 
-function UserList( {users, onRemove, onToggle} ) {
+function UserList( {users} ) {
   return (
     <div>
       {users.map( user => (
-        <User user={user} key={user.id} onRemove={onRemove} onToggle={onToggle}/>
+        <User user={user} key={user.id}/>
       ))}
     </div>
   );
